仲介会社一括仮押さえ編集より個別の仮押さえ削除完了

+++ app/Http/Controllers/Admin/MultiplesController.php
@@ -189,7 +189,7 @@ class MultiplesController extends Controller
 
   public function agent_edit($multiple_id, $venue_id)
   {
-    $multiple = MultipleReserve::find($multiple_id);
+    $multiple = MultipleReserve::with('pre_reservations.pre_enduser', 'pre_reservations.pre_bill')->find($multiple_id);
     $venue = Venue::find($venue_id);
     return view('admin.multiples.agent_edit', [
       'multiple' => $multiple,
@@ -368,6 +368,10 @@ class MultiplesController extends Controller
 
   public function SPDestroy(Request $request)
   {
+    if ($request->destroy_from_agent) {
+      return $this->destroy_from_agent($request);
+    }
+
     $shapeRequest = $request->except(['_method', '_token', 'multi_id']);
     $multiple = MultipleReserve::with('pre_reservations')->find($request->multi_id);
 
@@ -375,23 +379,52 @@ class MultiplesController extends Controller
       return redirect()->route('admin.multiples.show', $multiple->id)->with('flash_message_error', '仮押さえが選択されていません');
     }
 
-    DB::transaction(function () use ($shapeRequest) {
-      foreach ($shapeRequest as $value) {
-        $pre_reservation = PreReservation::find($value);
-        $pre_reservation->delete();
-      }
-    });
+    // DB::transaction(function () use ($shapeRequest) {
+    //   foreach ($shapeRequest as $value) {
+    //     $pre_reservation = PreReservation::find($value);
+    //     $pre_reservation->delete();
+    //   }
+    // });
 
-    $multiple = MultipleReserve::with('pre_reservations')->find($request->multi_id);
-    if (!empty($multiple->pre_reservations->toArray())) {
-      //まだ仮押さえがあるなら
-      return redirect()->route('admin.multiples.show', $multiple->id)->with('flash_message', '仮押さえの削除に成功しました。');
-    } else {
-      //仮押さえがない
-      DB::transaction(function () use ($multiple) {
-        $multiple->delete();
+    // $multiple = MultipleReserve::with('pre_reservations')->find($request->multi_id);
+    // if (!empty($multiple->pre_reservations->toArray())) {
+    //   //まだ仮押さえがあるなら
+    //   return redirect()->route('admin.multiples.show', $multiple->id)->with('flash_message', '仮押さえの削除に成功しました。');
+    // } else {
+    //   //仮押さえがない
+    //   DB::transaction(function () use ($multiple) {
+    //     $multiple->delete();
+    //   });
+    //   return redirect()->route('admin.multiples.index')->with('flash_message', '仮押さえの削除に成功しました。');
+    // }
+  }
+
+  public function destroy_from_agent($request)
+  {
+    if ($request->destroy_from_agent == 1) {
+      $shapeRequest = $request->except(['_method', '_token', 'multi_id', 'destroy_from_agent']);
+      $multiple = MultipleReserve::with('pre_reservations')->find($request->multi_id);
+      if (empty($shapeRequest)) {
+        return redirect()->route('admin.multiples.agent_show', $multiple->id)->with('flash_message_error', '仮押さえが選択されていません');
+      }
+      DB::transaction(function () use ($shapeRequest) {
+        foreach ($shapeRequest as $value) {
+          $pre_reservation = PreReservation::find($value);
+          $pre_reservation->delete();
+        }
       });
-      return redirect()->route('admin.multiples.index')->with('flash_message', '仮押さえの削除に成功しました。');
+
+      $multiple = MultipleReserve::with('pre_reservations')->find($request->multi_id);
+      if (!empty($multiple->pre_reservations->toArray())) {
+        //まだ仮押さえがあるなら
+        return redirect()->route('admin.multiples.agent_show', $multiple->id)->with('flash_message', '仮押さえの削除に成功しました。');
+      } else {
+        //仮押さえがない
+        DB::transaction(function () use ($multiple) {
+          $multiple->delete();
+        });
+        return redirect()->route('admin.multiples.index')->with('flash_message', '仮押さえの削除に成功しました。');
+      }
     }
   }
 }
